+++ app/(screens)/dashboard/_layout.tsx
@@ -5,26 +5,26 @@ import Header from '@/app/components/shared/Header'
 export default function TabLayout() {
   return (
     <>
-      <Header pageTitle="Dashboard" />
+      {/* <Header pageTitle="Dashboard" /> */}
       <Tabs screenOptions={{ headerShown: false }}>
         <Tabs.Screen
-          name="/dashboard/super-admin"
+          name='/dashboard/super-admin'
           options={{
             title: 'Admin',
           }}
         />
         <Tabs.Screen
-          name="/dashboard/teacher"
+          name='/dashboard/teacher'
           options={{
             title: 'Professor',
           }}
         />
         <Tabs.Screen
-          name="/dashboard/student"
+          name='/dashboard/student'
           options={{
             title: 'Estudante',
           }}
-          />
+        />
       </Tabs>
     </>
   )
